+++ src/data/duke.txt
@@ -8,4 +8,3 @@ event me /from 30-11-9999 2359/to 01-01-+10000 0000
 event me /from 01-01-+10000 0000/to 01-01-+10000 0000
 event me /from 30-11-9999 2359/to 01-01-+10000 0000
 event me /from 01-12-9999 0000/to 01-01-+10000 0000
-event me /from 31-12-9999 2359/to 01-01-+10000 0000
