+++ src/main/java/DukeExceptions/DukeException.java
@@ -1,0 +1,7 @@
+package main.java.DukeExceptions;
+
+public class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
